Rename input number validator
