+++ packages/perseus/src/widgets/input-number/input-number.test.ts
@@ -12,7 +12,7 @@ import {mockStrings} from "../../strings";
 import {renderQuestion} from "../__testutils__/renderQuestion";
 
 import InputNumber from "./input-number";
-import inputNumberValidator from "./input-number-validator";
+import scoreInputNumber from "./score-input-number";
 import {question3 as question} from "./input-number.testdata";
 
 import type {
@@ -250,7 +250,7 @@ describe("invalid", function () {
     });
 
     it("should handle invalid answers with no error callback", function () {
-        const err = inputNumberValidator(
+        const err = scoreInputNumber(
             {currentValue: "x+1"},
             options,
             mockStrings,
